+++ src/triangulation.pyx
@@ -153,10 +153,10 @@ cpdef list vpoints_configure(object vpoints_, object inputs, dict status = None)
                     continue
                 links.update(vpoint_.links)
                 vpoints[node] = VPoint.c_slider_joint(
-                    ",".join([vpoint.links[0]] + [
+                    [vpoint.links[0]] + [
                         link_ for link_ in vpoint_.links
                         if (link_ not in vpoint.links)
-                    ]),
+                    ],
                     VJoint.RP,
                     vpoint.angle,
                     vpoint_.x,